Add files via upload
# Covid SQL database project

In this project, we will create a SQL database for Covid-19 data. The database will contain tables for tracking cases, deaths, and recoveries from Covid-19. We will also create tables for tracking demographics, such as age and gender, and locations, such as states or countries.

## Tables

### Cases

The cases table will track the number of confirmed cases of Covid-19. The table will include the following columns:

- `case_id`: a unique identifier for each case
- `date`: the date the case was confirmed
- `location_id`: a foreign key linking to the location table
- `age`: the age of the person who tested positive
- `gender`: the gender of the person who tested positive

### Deaths

The deaths table will track the number of deaths due to Covid-19. The table will include the following columns:

- `death_id`: a unique identifier for each death
- `date`: the date the death occurred
- `location_id`: a foreign key linking to the location table
- `age`: the age of the person who died
- `gender`: the gender of the person who died

### Recoveries

The recoveries table will track the number of people who have recovered from Covid-19. The table will include the following columns:

- `recovery_id`: a unique identifier for each recovery
- `date`: the date the recovery was confirmed
- `location_id`: a foreign key linking to the location table
- `age`: the age of the person who recovered
- `gender`: the gender of the person who recovered

### Locations

The locations table will track the different locations where Covid-19 data is being collected. The table will include the following columns:

- `location_id`: a unique identifier for each location
- `location_name`: the name of the location (e.g. state, country, or city)

## Conclusion

By creating a SQL database for Covid-19 data, we can easily track and analyze the spread of the virus. This database can be used by public health officials, researchers, and others to make informed decisions and take appropriate actions to mitigate the impact of Covid-19.